refactor: reorganize teacher controllers and routes; remove unused files and update references
