+++ routes/teachers.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const router = express.Router();
-const teacherDashboardController = require("../controllers/teacher/dashboard");
-const teacherStudentController = require("../controllers/teacher/studentsController");
-const teacherLessonController = require("../controllers/teacher/lessonController");
+const teacherDashboardController = require("../controllers/teachers/dashboard");
+const teacherStudentController = require("../controllers/teachers/studentsController");
+const teacherLessonController = require("../controllers/teachers/lessonController");
 const { ensureAuth } = require("../middleware/auth");
 const upload = require("../middleware/multer");
 
@@ -57,13 +57,13 @@ router.delete(
 
 // display new lesson form
 router.get(
-    "/students/:studentId/lesson/new",
+    "/students/:studentId/lessons/new",
     ensureAuth,
     teacherLessonController.getNewLessonForm
 );
 // // create a new lesson plan
 router.post(
-    "/students/:studentId/lesson",
+    "/students/:studentId/lessons",
     ensureAuth,
     upload.single("file"),
     // todo: create a new lesson plan as a teacher
@@ -71,25 +71,25 @@ router.post(
 );
 
 // router.get(
-//     "/students/:studentId/lesson/:lessonId",
+//     "/students/:studentId/lessons/:lessonId",
 //     ensureAuth,
 //     teacherController.getLesson
 // );
 // router.get(
-//     "/students/:studentId/lesson/:lessonId/edit",
+//     "/students/:studentId/lessons/:lessonId/edit",
 //     ensureAuth,
 //     teacherController.getEditLessonForm
 // );
 
 // router.put(
-//     "/students/:studentId/lesson/:lessonId",
+//     "/students/:studentId/lessons/:lessonId",
 //     ensureAuth,
 //     teacherController.updateLesson
 // );
 
 // // delete a lesson plan
 // router.delete(
-//     "/students/:studentId/lesson/:lessonPlanId",
+//     "/students/:studentId/lessons/:lessonPlanId",
 //     ensureAuth,
 //     // todo: controller to delete a lesson plan as a teacher
 //     teacherController.deleteLesson
