+++ views/teachers/lessons/new.ejs
@@ -1,4 +1,4 @@
-<%- include('../partials/header') -%>
+<%- include('../../partials/header') -%>
 <div class="container">
     <div class="row mt-5">
         <div class="col-6">
@@ -50,4 +50,4 @@
         </div>
     </div>
 </div>
-<%- include('../partials/footer') -%>
+<%- include('../../partials/footer') -%>
